% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacEDA-package.R
\docType{package}
\name{qacEDA-package}
\alias{qacEDA}
\alias{qacEDA-package}
\title{qacEDA: Functions to Facilitate Exploratory Data Analysis}
\description{
This package provides functions for descriptive statistics,
  data management, and data visualization.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkabacoff/qacEDA}
  \item Report bugs at \url{https://github.com/rkabacoff/qacEDA/issues}
}

}
\author{
\strong{Maintainer}: QAC 356 \email{rkabacoff@wesleyan.edu}

}
\keyword{internal}
